Add files via upload
In this paper, is presented an approach on Feature Engineering for applications on Machine
learning in Remote Sensing. Singular Value Decomposition and Principal Component
Analysis were applied over a group of bands in order to reduce the amount of data and
dimensionality of it. The results provided demonstrate the relevance of Feature Engineering
in the field of computer sciences - ML, as it improves drastically the results of a model
without changing the parameters of the model or the characteristics of the data-set.